requirements.txt
flask
python-dotenv
flask_sqlalchemy
flask_bcrypt
flask_migrate
paho-mqtt